% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cant_solve.R
\name{cant.solve}
\alias{cant.solve}
\title{Test if you can invert a matrix}
\usage{
cant.solve(m)
}
\arguments{
\item{t}{a matrix}
}
\value{
True or false
}
\description{
Test if you can invert a matrix
}

