% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cragg_errors1.R
\name{cragg_errs}
\alias{cragg_errs}
\title{Resample for truncated normal errors}
\usage{
cragg_errs(cov, pi, x1, gamma, beta, n, z)
}
\arguments{
\item{cov}{the covariance matrix. This should be untransformed, the
terms will be multiplied by the coefficients within the resampling
procedure.}

\item{pi}{a vector of coefficients for the first stage regression}

\item{x1}{your exogenous variables (a dataframe)}

\item{gamma}{a vector of coefficients for the second stage probit}

\item{beta}{a vector of coefficients for the second stage linear regression}

\item{n}{the number of errors to be generated}

\item{z}{your instrument (a dataframe)}
}
\value{
returns a list of starting parameters for the coefficients
and the covariance matrix
}
\description{
The truncated multivariate normal errors can be generated
in one of two ways. Either we sample the first stage and probit errors
first and then resample the linear regression's errors until we have
positive values, or we resample the vector of all three errors until
the linear value is positive. This is the second method, use cragg_errs2
for the second, which is less intuitive.
}

