% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cragg_errors3.R
\name{cragg_errs3}
\alias{cragg_errs3}
\title{Resample for truncated normal errors}
\usage{
cragg_errs3(cov, pi, x1, gamma, beta, n, z)
}
\arguments{
\item{cov}{the covariance matrix. This should be untransformed, the
terms will be multiplied by the coefficients within the resampling
procedure.}

\item{pi}{a vector of coefficients for the first stage regression}

\item{x1}{your exogenous variables (a dataframe)}

\item{gamma}{a vector of coefficients for the second stage probit}

\item{beta}{a vector of coefficients for the second stage linear regression}

\item{n}{the number of errors to be generated}

\item{z}{your instrument (a dataframe)}
}
\value{
returns a list of your errors and the three generated variables:
the endogenous regressor, the censoring variable and the outcome variable
}
\description{
The truncated multivariate normal errors can be generated
in one of three ways. 1) we sample the first stage and probit errors
first and then resample the linear regression's errors until we have
positive values. 2) we resample the vector of all three errors until
the linear value is positive. 3) we send all our variables to zero first
then resample the full vector, accepting if it lies in Q1.
This is the third method.
}
