% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cragg_errors_MG.R
\name{cragg_errs_MG}
\alias{cragg_errs_MG}
\title{Resample for truncated normal errors}
\usage{
cragg_errs_MG(cov, pi, x1, gamma, beta, n, z)
}
\arguments{
\item{cov}{the covariance matrix. This should be untransformed, the
terms will be multiplied by the coefficients within the resampling
procedure.}

\item{pi}{a vector of coefficients for the first stage regression}

\item{x1}{your exogenous variables (a dataframe)}

\item{gamma}{a vector of coefficients for the second stage probit}

\item{beta}{a vector of coefficients for the second stage linear regression}

\item{n}{the number of errors to be generated}

\item{z}{your instrument (a dataframe)}
}
\value{
returns a list of your errors and the three generated variables:
the endogenous regressor, the censoring variable and the outcome variable
}
\description{
There are a bunch of ways of thinking about this DGP. This is
what Max G'Sell had in mind. We generate x2 and y0 first, and then sample
from y1 conditionally on those two and on itself being positive.
This can only handle one endogenous variable for now. Testing it with
more than one exogenous variable.
}
