% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleIVsim.R
\name{hurdle.IV.sim}
\alias{hurdle.IV.sim}
\title{Simulate a dataset}
\usage{
hurdle.IV.sim(formula = F, pi = c(1, -1, 3), gamma = c(-0.2, 0.8, 0.07),
  beta = c(0.05, 0.06, 0.02), endog_reg = list(), exog_mean = 1,
  exog_sd = 1, z_mean = 3, z_sd = 1, endog_sd = 5, y_sd = 2,
  rho = 0.2, tau0 = 0.3, tau1 = 0.1, n = 10000, options = list(silent
  = F, cragg_errors = 1), type = "lognormal")
}
\arguments{
\item{formula}{For now, must be False. In the future, will allow you to leave
out exogenous variables you use in first stage regressions from second
stage regressions.}

\item{pi}{a vector of the coefficients on your first stage regression.
Should be in the order: intercept, exogenous variables, instruments.
Defaults to (1,-1,3).}

\item{gamma}{a vector of the coefficients on your second stage probit
regression. Should be in the order: intercept, exogenous variables,
endogenous variable. Defaults to (-.2,.8,.07)}

\item{beta}{a vector of the coefficients on your second stage linear
regression. Should be in the order: intercept, exogenous variables,
endogenous variable. Defaults to (.05,.06,.02)}

\item{endog_reg}{for now, leave as an empty list. Will create one endogenous
regression which includes the instrument and all exogenous variables}

\item{exog_mean}{the mean(s) of your exogenous variable(s). Defaults to 1}

\item{exog_sd}{the standard deviation(s) of your exogenous variable(s).
Defaults to 1.}

\item{z_mean}{the mean(s) of your instrument(s). Defaults to 3.}

\item{z_sd}{the standard deviation(s) of your instrument(s).
Defaults to 1.}

\item{y_sd}{the standard deviation of your second stage linear regression.
Defaults to 2.}

\item{rho}{the covariance term between second stage linear and probit
regressions. Defaults to 0.2.}

\item{tau0}{the covariance term between the second stage probit and the
first stage regression. Defaults to 0.3.}

\item{tau1}{the covariance term between the second stage linear regression and the
first stage regression. Defaults to 0.1.}

\item{n}{the number of observations. Defaults to 10,000.}

\item{options}{Options are: silent = F (will print histograms and percent censored),
cragg_errors = {1,2} cragg error type 1 resamples entire vector of errors.
See documentation for cragg_errs for more detail.}

\item{type}{defaults to lognormal. Enter "cragg" for cragg model.}
}
\value{
returns a simulated dataset
}
\description{
Simulates a dataset from either Cragg or Lognormal distributions.
Allows you to specify the data generating process, though it still
can't handle more than one endogenous regression
}

