% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cov_baby.R
\name{make.cov}
\alias{make.cov}
\title{Helper function to make covariance matrix}
\usage{
make.cov(rho, tau0, tau1, y_sd, endog_sd)
}
\arguments{
\item{rho}{covariance between second stage regressions}

\item{tau0}{covariance between second stage probit and first stage
regression}

\item{tau1}{covariance between second stage linear and first stage
regression}

\item{y_sd}{second stage linear regression standard deviation}

\item{endog_sd}{first stage linear regression standard deviation}
}
\value{
returns a covariancne matrix
}
\description{
Specify elements of the covariance matrix and it organizes
them for you
}

