% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cov.R
\name{make.covTrans}
\alias{make.covTrans}
\title{Make a proper covariance matrix}
\usage{
make.covTrans(a, num_endog, gamma, beta, option = "mat", noname = F)
}
\arguments{
\item{a}{can be a matrix or vector of what should be included in 
the covariance matrix}

\item{num_endog}{the number of endogenous regressors}

\item{gamma}{a vector of your second stage probit regression parameters}

\item{beta}{a vector of your second stage linear regression parameters}

\item{option}{either "mat" if a is a matrix, "vector" if a is all the
elements of the matrix but in vector form or "parameters" in which case
a should be a list of the important parameters:
a = list(rho,tau0,tau1,y_sd,endog_sd)}
}
\value{
returns the covariance matrix
}
\description{
In these regressions, we need to transform the covariance
matrix using the coefficients in our regressions. This function produces that 
covariance matrix.
}

