% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_names.R
\name{make.names}
\alias{make.names}
\title{Helper fn}
\usage{
make.names(len, pref)
}
\arguments{
\item{len}{the length of your name vector}

\item{pref}{the prefix}
}
\value{
returns a vector of names of the form x1, x2, .... xN
}
\description{
Names the columns of your dataset to help out
}

