% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_pieces.R
\name{name.pieces}
\alias{name.pieces}
\title{Name the elements that go into your regression function.}
\usage{
name.pieces(t, a1 = len_cov, a2 = num_betas, a3 = num_endog,
  a4 = numpis)
}
\arguments{
\item{t}{a vector which should start with the covariates, then the
betas, then the gammas, then the pis}

\item{a1}{the number of elements in the covariance matrix}

\item{a2}{the number beta coefficients (2nd stage linear regression)}

\item{a3}{the number endogenous regressors}

\item{a4}{a list of the number of pi coefficients (1st stage regression)
for each endogenous regressor}
}
\value{
The vector renamed according to what it needs to be to work in the
likelihood functions
}
\description{
There is likely
a cleaner way to do this, but it allows for more flexibility as to
how many of each parameter you're going to have.
}
