% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstitute_cov.R
\name{reconstitute.cov}
\alias{reconstitute.cov}
\title{Reconstitute a covariance matrix}
\usage{
reconstitute.cov(vals, num, chol = myChol)
}
\arguments{
\item{vals}{a vector of covariance matrix elements}

\item{num}{the number of endogenous regressors}

\item{chol}{True/False. If true, will re-transform to original}
}
\value{
The vector renamed according to what it needs to be to work in the 
likelihood functions
}
\description{
This takes a vector and reconstitutes it into the matrix
you want in the end. The first term is always 1 under the assumptions
of the IV hurdle models. If you did a cholesky decomposition it will 
un-do that tranformation for you.
}

