% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_fn.R
\name{start.val}
\alias{start.val}
\title{Get the starting points for your regression}
\usage{
\method{start}{val}(formula, endog_reg, data, type)
}
\arguments{
\item{formula}{the second stage linear regression (outcome ~ exogenous + endogenous)}

\item{endog_reg}{a list of endogenous regressions. By default it will 
include all exogenous regressors and one instrument}

\item{data}{your data, as a dataset}

\item{type}{either "lognormal" or "cragg"}
}
\value{
returns a list of starting parameters for the coefficients
and the covariance matrix
}
\description{
The maximum likelihood procedure needs a place to start.
This runs traditional IV probit/linear regressions to get starting points.
}

